+++ dev/user.clj
@@ -14,7 +14,7 @@
   (b/time+ (b/tarai 10 5 3))
   (b/quick (b/tarai 10 5 3))
 
-  (d/start "target/db.sqlite")
+  (d/start "storage/db.sqlite")
   (d/conn?)
 
   (d/puts [{:db/add -1, :name "hiroshi", :age 63, :like "clojure"}])
@@ -35,8 +35,21 @@
 
   (d/stop)
 
-  (d/start "target/db.sqlite")
+  (d/start "storage/db.sqlite")
+
   (d/conn?)
+
+  (d/q '[:find ?name ?age ?like
+         :where
+         [?e :name ?name]
+         [?e :age ?age]
+         [?e :like ?like]])
+
+  (d/q '[:find ?name ?database
+         :where
+         [?e :name ?name]
+         [?e :database ?database]])
+
   (d/stop)
 
   :rcf)