+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 ## 0.3.0 (2025-04-19)
 
 * (start) ... on-memory mode
-* (start db) ... backend sqlite mode
+* (start "db") ... backend sqlite mode. "db" is the path of sqlite db saved.
 
 ## 0.2.5 (2025-04-19)